Conflicts:
	actor-sdk/sdk-core/core/core-shared/src/main/java/im/actor/core/modules/api/SequenceActor.java

+++ actor-sdk/sdk-core/core/core-shared/src/main/java/im/actor/core/api/rpc/RequestLoadEphermalPublicKeys.java
@@ -5,34 +5,27 @@ package im.actor.core.api.rpc;
 
 import im.actor.runtime.bser.*;
 import im.actor.runtime.collections.*;
-
 import static im.actor.runtime.bser.Utils.*;
-
 import im.actor.core.network.parser.*;
-
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.annotations.NotNull;
-
 import com.google.j2objc.annotations.ObjectiveCName;
-
 import java.io.IOException;
 import java.util.List;
 import java.util.ArrayList;
-
 import im.actor.core.api.*;
 
 public class RequestLoadEphermalPublicKeys extends Request<ResponsePublicKeys> {
 
     public static final int HEADER = 0xa2b;
-
     public static RequestLoadEphermalPublicKeys fromBytes(byte[] data) throws IOException {
         return Bser.parse(new RequestLoadEphermalPublicKeys(), data);
     }
 
     private ApiUserOutPeer userPeer;
-    private long keyGroupId;
+    private int keyGroupId;
 
-    public RequestLoadEphermalPublicKeys(@NotNull ApiUserOutPeer userPeer, long keyGroupId) {
+    public RequestLoadEphermalPublicKeys(@NotNull ApiUserOutPeer userPeer, int keyGroupId) {
         this.userPeer = userPeer;
         this.keyGroupId = keyGroupId;
     }
@@ -46,14 +39,14 @@ public class RequestLoadEphermalPublicKeys extends Request<ResponsePublicKeys> {
         return this.userPeer;
     }
 
-    public long getKeyGroupId() {
+    public int getKeyGroupId() {
         return this.keyGroupId;
     }
 
     @Override
     public void parse(BserValues values) throws IOException {
         this.userPeer = values.getObj(1, new ApiUserOutPeer());
-        this.keyGroupId = values.getLong(2);
+        this.keyGroupId = values.getInt(2);
     }
 
     @Override
@@ -62,7 +55,7 @@ public class RequestLoadEphermalPublicKeys extends Request<ResponsePublicKeys> {
             throw new IOException();
         }
         writer.writeObject(1, this.userPeer);
-        writer.writeLong(2, this.keyGroupId);
+        writer.writeInt(2, this.keyGroupId);
     }
 
     @Override
